FindBugs said there is a bug!

+++ game-server/src/com/aionemu/gameserver/services/HousingBidService.java
@@ -482,7 +482,7 @@ public class HousingBidService extends AbstractCronTask {
 				// make the new house inactive until the old one is sold
 				obtainedHouse.setStatus(HouseStatus.INACTIVE);
 				result = AuctionResult.GRACE_START;
-				time = new DateTime(getRunTime() * 1000).plusWeeks(2).getMillis();
+				time = new DateTime((long) getRunTime() * 1000).plusWeeks(2).getMillis();
 			}
 		}
 		obtainedHouse.setOwnerId(winner.getPlayerObjId());
